requirements.txt
pandas
numpy
matplotlib
python-dotenv
flask